Set $delete_printers = 0; by default and added warning to test thoroughly before deploying
While it should be relatively safe to turn on, if the $print_server string is found in the URI path of a printer not connected to $print_server, it will likely be deleted from the client computer when the script is run.  Using the full domain name for $print_server is a good way to be relatively sure it's safe to turn the $delete_printers setting on.